# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "API Cantina da Sandra",
    "description": "API para o sistema de pedidos da cantina da Sandra"
  },
  "host": "localhost:4000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "scheme": "bearer",
      "bearerFormat": "JWT"
    }
  },
  "paths": {
...
    "/pedidos/": {
      "get": {
        "tags": [
          "Pedidos"
        ],
        "description": "Endpoint para obter todos os pedidos de um cliente.",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Token de autenticação",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de pedidos",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "Id": {
                    "type": "string",
                    "example": "number"
                  },
                  "Valor": {
                    "type": "string",
                    "example": "number"
                  },
                  "Data": {
                    "type": "string",
                    "example": "Date"
                  },
                  "MetPag": {
                    "type": "string",
                    "example": "string"
                  },
                  "ClienteId": {
                    "type": "string",
                    "example": "number"
                  },
                  "Clientes": {
                    "type": "object",
                    "properties": {
                      "Id": {
                        "type": "string",
                        "example": "number"
                      },
                      "Nome": {
                        "type": "string",
                        "example": "string"
                      },
                      "Login": {
                        "type": "string",
                        "example": "string"
                      },
                      "Senha": {
                        "type": "string",
                        "example": "string"
                      },
                      "Ativado": {
                        "type": "string",
                        "example": "boolean"
                      },
                      "CPF": {
                        "type": "string",
                        "example": "string"
                      }
                    }
                  },
                  "ProdutosPedidos": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "Id": {
                          "type": "string",
                          "example": "number"
                        },
                        "ProdutoId": {
                          "type": "string",
                          "example": "number"
                        },
                        "PedidoId": {
                          "type": "string",
                          "example": "number"
                        },
                        "Quantidade": {
                          "type": "string",
                          "example": "number"
                        },
                        "Produtos": {
                          "type": "object",
                          "properties": {
                            "Id": {
                              "type": "string",
                              "example": "number"
                            },
                            "Nome": {
                              "type": "string",
                              "example": "string"
                            },
                            "Descricao": {
                              "type": "string",
                              "example": "string"
                            },
                            "Preco": {
                              "type": "string",
                              "example": "number"
                            },
                            "Categoria": {
                              "type": "string",
                              "example": "string"
                            },
                            "URLImagem": {
                              "type": "string",
                              "example": "string"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "Pedido não encontrado"
          }
        }
      }
    },
    "/pedidos/{id}": {
      "get": {
        "tags": [
          "Pedidos"
        ],
        "description": "Endpoint para obter um pedido.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "number",
            "description": "Id do pedido"
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "Token de autenticação",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Pedido",
            "schema": {
              "type": "object",
              "properties": {
                "Id": {
                  "type": "string",
                  "example": "number"
                },
                "Valor": {
                  "type": "string",
                  "example": "number"
                },
                "Data": {
                  "type": "string",
                  "example": "Date"
                },
                "MetPag": {
                  "type": "string",
                  "example": "string"
                },
                "ClienteId": {
                  "type": "string",
                  "example": "number"
                },
                "Clientes": {
                  "type": "object",
                  "properties": {
                    "Id": {
                      "type": "string",
                      "example": "number"
                    },
                    "Nome": {
                      "type": "string",
                      "example": "string"
                    },
                    "Login": {
                      "type": "string",
                      "example": "string"
                    },
                    "Senha": {
                      "type": "string",
                      "example": "string"
                    },
                    "Ativado": {
                      "type": "string",
                      "example": "boolean"
                    },
                    "CPF": {
                      "type": "string",
                      "example": "string"
                    }
                  }
                },
                "ProdutosPedidos": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "Id": {
                        "type": "string",
                        "example": "number"
                      },
                      "ProdutoId": {
                        "type": "string",
                        "example": "number"
                      },
                      "PedidoId": {
                        "type": "string",
                        "example": "number"
                      },
                      "Quantidade": {
                        "type": "string",
                        "example": "number"
                      },
                      "Produtos": {
                        "type": "object",
                        "properties": {
                          "Id": {
                            "type": "string",
                            "example": "number"
                          },
                          "Nome": {
                            "type": "string",
                            "example": "string"
                          },
                          "Descricao": {
                            "type": "string",
                            "example": "string"
                          },
                          "Preco": {
                            "type": "string",
                            "example": "number"
                          },
                          "Categoria": {
                            "type": "string",
                            "example": "string"
                          },
                          "URLImagem": {
                            "type": "string",
                            "example": "string"
                          }
                        }
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "Pedido não encontrado"
          }
        }
      },
      "put": {
        "tags": [
          "Pedidos"
        ],
        "description": "Endpoint para atualizar um pedido.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "number",
            "description": "Id do pedido"
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "Token de autenticação",
            "type": "string"
          },
          {
            "name": "pedido",
            "in": "body",
            "description": "Pedido",
            "schema": {
              "type": "object",
              "properties": {
                "Valor": {
                  "type": "string",
                  "example": "number"
                },
                "MetPag": {
                  "type": "string",
                  "example": "string"
                },
                "ProdutosPedidos": {
                  "type": "string",
                  "example": "Produto[]"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pedido atualizado",
            "schema": {
              "type": "object",
              "properties": {
                "Id": {
                  "type": "string",
                  "example": "number"
                },
                "Valor": {
                  "type": "string",
                  "example": "number"
                },
                "Data": {
                  "type": "string",
                  "example": "Date"
                },
                "MetPag": {
                  "type": "string",
                  "example": "string"
                },
                "ClienteId": {
                  "type": "string",
                  "example": "number"
                },
                "ProdutosPedidos": {
                  "type": "string",
                  "example": "Produto[]"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Pedido não encontrado"
          },
          "500": {
            "description": "Erro ao atualizar pedido"
          }
        }
      }
    },
    "/pedidos/adicionar": {
      "post": {
        "tags": [
          "Pedidos"
        ],
        "description": "Endpoint para adicionar um pedido.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "Token de autenticação",
            "type": "string"
          },
          {
            "name": "pedido",
            "in": "body",
            "description": "Pedido",
            "schema": {
              "type": "object",
              "properties": {
                "Valor": {
                  "type": "string",
                  "example": "number"
                },
                "MetPag": {
                  "type": "string",
                  "example": "string"
                },
                "ProdutosPedidos": {
                  "type": "string",
                  "example": "Produto[]"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pedido adicionado",
            "schema": {
              "type": "object",
              "properties": {
                "Id": {
                  "type": "string",
                  "example": "number"
                },
                "Valor": {
                  "type": "string",
                  "example": "number"
                },
                "Data": {
                  "type": "string",
                  "example": "Date"
                },
                "MetPag": {
                  "type": "string",
                  "example": "string"
                },
                "ClienteId": {
                  "type": "string",
                  "example": "number"
                },
                "ProdutosPedidos": {
                  "type": "string",
                  "example": "Produto[]"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Erro ao adicionar pedido"
          }
        }
      }
    },
...
  },
...
}